Fix warning of open_inode() failure
open_inode() warns if finding inode is not available in itree.

Usually, finding inode should be available in itree though. The bitmap
inode can be unavailable after mkfs, because bitmap inode is only
written after unify.

So, lift up warning to caller of open_inode(), and warn if finding
inode should be available.

Signed-off-by: OGAWA Hirofumi <hirofumi@mail.parknet.co.jp>