Choose and assign inum after creating dirent
Refactoring towards choosing inode after creating directory entry.
If tux_create_dirent sees an inode with TUX_INVALID_INO then it knows this
is an inode create and it must choose the inode number. This is a slightly
obscure way to do things, so insert a paranoia check in the create_link
caller, where it is not completely obvious that the inode always has a valid
inode number.

At this point the inode choosing is done after the directory entry create, so
we have the opportunity to use directory entry position as part of the inode
number selection heuristics.

Signed-off-by: Daniel Phillips <d.phillips@partner.samsung.com>
Signed-off-by: OGAWA Hirofumi <hirofumi@mail.parknet.co.jp>