Fix multiple tests on same state in tests/commit.c
For example, test05 in tests/commit.c has test05.1 and test05.2. But
test05.1 changes on-disk data, not only in-core.

So, changes in test05.1 is visible to test05.2. But test05.2 is
assuming on-disk state before test05.1.

To fix this bugs, this patch adds ability to snapshot of current
on-disk data, and restore.

Signed-off-by: OGAWA Hirofumi <hirofumi@mail.parknet.co.jp>