Introduce direct extent as alternative to btree root
For a relatively small file that can be allocated linearly without
holes the btree root is overloaded as a direct extent pointer. On
media, the high bit of the btree depth is used to indicate that
the root is actually a direct extent. This takes advantage of the
fact that the data paths for direct extent and btree root are
nearly the same. Whether this is the best approach is a good
question. We should try this, see how it performs, then consider
the ideal solution.

Signed-off-by: Daniel Phillips <d.phillips@partner.samsung.com>
Signed-off-by: OGAWA Hirofumi <hirofumi@mail.parknet.co.jp>