Make block allocation goal depend on inode number and logical file offset
Make block allocation goal depend on inode number and logical file offset.
If the inode is outside the volume block range, "fold" it by masking against
the volume mask, which is the smallest binary number greater than the volume
size. If still outside, subtract half the volume mask so that it is inside.
Volume coverage of masked values is not perfectly even, but close enough.
The purpose of masking this way is to resist fragmentation if the volume size
changes.

Use linear allocation without extrapolation when inodes are being written in
sequential order. If this is too far from the extrapolated goal, then use the
extrapolated goal instead.

Signed-off-by: Daniel Phillips <d.phillips@partner.samsung.com>
Signed-off-by: OGAWA Hirofumi <hirofumi@mail.parknet.co.jp>